style: Move Horizontal and Vertical Bars
