+++ frontend/src/components/desktop_apps/AppWindow.tsx
@@ -5,7 +5,7 @@ import BitMinerUpgradeScreen from "./screens/BitMinerAppScreen";
 import LearnAppScreen from "./screens/LearnAppScreen";
 import LoginContainer from "../auth/LoginContainer";
 import SettingsAppScreen from "./screens/SettingsAppScreen";
-import TitleBar from "./screens/TitleBar";
+import TitleBar from "./TitleBar";
 import PuzzleAppScreen from "./screens/PuzzleAppScreen";
 import HelpAppScreen from "./screens/HelpAppScreen";
 import TerminalAppScreen from "./screens/TermialAppScreen";
@@ -17,8 +17,24 @@ interface AppWindowProps {
   setOpen: (at: AppType) => void;
 }
 
-const AppWindow: React.FC<AppWindowProps> = ({ open, setOpen }) => {
+export default function AppWindow({ open, setOpen }: AppWindowProps) {
   const windowBackgroundColor = useSelector((state: IGameState) => state.styleData.backgroundColor.window);
+  let style: React.CSSProperties = {
+    width: "80%",
+    height: "80%",
+    background: windowBackgroundColor,
+    display: "flex",
+    alignItems: "center",
+    flexDirection: "column",
+    justifyContent: "flex-start",
+    color: "black",
+    position: "absolute",
+    zIndex: "1",
+    left: "10%",
+    top: "10%",
+    border: "1px solid gray"
+  };
+
   let appWindowElements = null;
   switch (open) {
     case AppType.Collector:
@@ -26,8 +42,6 @@ const AppWindow: React.FC<AppWindowProps> = ({ open, setOpen }) => {
       break;
     case AppType.Upgrades:
       appWindowElements = <UpgradeAppScreen></UpgradeAppScreen>;
-      break;
-    case AppType.Store:
       break;
     case AppType.Login:
       appWindowElements = <LoginContainer></LoginContainer>;
@@ -41,56 +55,21 @@ const AppWindow: React.FC<AppWindowProps> = ({ open, setOpen }) => {
     case AppType.Puzzle:
       appWindowElements = <PuzzleAppScreen></PuzzleAppScreen>;
       break;
-    case AppType.Help:
-      appWindowElements = <HelpAppScreen></HelpAppScreen>;
-      return (
-        <div
-          style={{
-            width: "40%",
-            height: "78%",
-            background: windowBackgroundColor,
-            display: "flex",
-            alignItems: "center",
-            flexDirection: "column",
-            justifyContent: "flex-start",
-            color: "white",
-            position: "absolute",
-            zIndex: "1",
-            left: "54%",
-            top: "5%",
-            border: "1px solid gray"
-          }}
-        >
-          <TitleBar open={open} setOpen={() => setOpen(null)}></TitleBar>
-          {appWindowElements}
-        </div>
-      );
     case AppType.Terminal:
       appWindowElements = <TerminalAppScreen></TerminalAppScreen>;
+      break;
+    case AppType.Help:
+      appWindowElements = <HelpAppScreen></HelpAppScreen>;
+      style = { ...style, width: "40%", height: "78%", color: "white", left: "54%", top: "5%" };
+      break;
+    case AppType.Store:
+      break;
   }
 
   return (
-    <div
-      style={{
-        width: "80%",
-        height: "80%",
-        background: windowBackgroundColor,
-        display: "flex",
-        alignItems: "center",
-        flexDirection: "column",
-        justifyContent: "flex-start",
-        color: "black",
-        position: "absolute",
-        zIndex: "1",
-        left: "10%",
-        top: "10%",
-        border: "1px solid gray"
-      }}
-    >
+    <div style={style}>
       <TitleBar open={open} setOpen={() => setOpen(null)}></TitleBar>
       {appWindowElements}
     </div>
   );
-};
-
-export default AppWindow;
+}
